feat(metadata): Add ability parse metadata using parsers;Refactor bom.py to ovoid import conflicts;
